State machine updated
So far only the "StateSetFixTrackDeg" state is implemented. Also it should be named smth better, like "SetParams", as we are setting all the params of a memory slot here. Also an eeprom should be integrated into it and write the on exit so we save this all to the eeprom